requirements.txt
rich
termcolor
mss
pyautogui
pyperclip
ollama
pywin32
pillow
rapidocr-onnxruntime
